requirements.txt
fastapi==0.111.0
uvicorn==0.23.2
opencv-python-headless==4.8.1.78
ultralytics==8.1.115
reportlab==4.1.3
Pillow==10.1.0
numpy==1.27.4
